Implemented the basics of a Turing Machine where the SOMs learn the perception/action rules and then we execute the Turing Machine right out of the manifolds. The tape and state variable are continuous, but in certain places we quantize for printing out or doing a discrete action like moving the tape head (a discrete amount).
The insight as using the BMU of the perception SOM as the location where the
Action SOM learned how to poke the world. This built an association map that
computes a set theoretic relation. I believe I'm hovering around the notion
that this specific construction of manifolds is turing complete (under certain
noise constraints), but I still need a choice model to finish it.