I wrote this path set a LONG time ago, but only now just commiting it.
Implement a turing machine example. The SOM learns the perception->action
vectors using an association map built with SOMs and then executes the
program which counts in binary.

Implement a read-only file system whose storage in embedded into the SOMs.

There is a small graphical foible when drawing the turing SOMs that I
should figure out.

There is no documentation, but when running the visualizer:
q exts
a is color of first 3 components of SOM vector
s is a quality map of the SOM. bright means similar, dark non-similar.